requirements.txt
# requirements.txt
# Core Python packages
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# Database
mysql-connector-python==8.1.0
SQLAlchemy==2.0.19

# Machine Learning
scikit-learn==1.3.0
prophet==1.1.4

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0

# Data Processing
requests==2.31.0
beautifulsoup4==4.12.2

# Model Persistence
joblib==1.3.1

# Visualization (optional for backend analysis)
plotly==5.15.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3